dxf: Center object in (0, 0)
Since most cad usages will have pieces centered in the origin axis and
only one outline is exported in the dxf, it makes sense to also center
the outline in the axis.

This works by introducing a modification centerOnOrigin that computes a
bounding box and translates polygon points according to its center.